Rename classes directory to lib
