+++ lib/monster.rb
